+++ test/data/safety/content/test/pin-with-require.js
@@ -1,5 +1,0 @@
-const logger = require("../../../../../lib/logger.js");
-
-process.env.NODE_CONFIG_ENV = "test";
-
-module.exports = logger;